refactor: template 폴더 index.js파일 분리
authPageTemplate.js
adminPageTemplate.js
